requirements.txt
ply==3.11
pillow~=11.0.0